--- a/.gitignore
+++ b/.gitignore
@@ -11,3 +11,4 @@ release/
 .claude/
 CLAUDE.md
 PLAN.md
+KEYS_MANAGEMENT.md