--- a/src/arbitrage/detector.rs
+++ b/src/arbitrage/detector.rs
@@ -331,6 +331,10 @@ mod tests {
             threshold: 1.0, // 1% threshold
             max_price_age_ms: 5000,
             rpc_url: None,
+            helius_api_key: None,
+            quicknode_api_key: None,
+            alchemy_api_key: None,
+            genesisgo_api_key: None,
             output_format: crate::output::OutputFormat::Table,
             min_price: 1.0,
             max_price: 10000.0,
@@ -436,6 +440,10 @@ mod tests {
             threshold: 0.01, // Very low threshold (0.01%)
             max_price_age_ms: 5000,
             rpc_url: None,
+            helius_api_key: None,
+            quicknode_api_key: None,
+            alchemy_api_key: None,
+            genesisgo_api_key: None,
             output_format: crate::output::OutputFormat::Table,
             min_price: 1.0,
             max_price: 10000.0,